add mantine
