+++ package.json
@@ -12,7 +12,11 @@
     "deploy": "gh-pages -d build"
   },
   "dependencies": {
+    "@emotion/react": "^11.13.3",
+    "@mantine/core": "^7.14.1",
+    "@mantine/hooks": "^7.14.1",
     "@radix-ui/react-icons": "^1.3.0",
+    "@tabler/icons-react": "^3.22.0",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "gray-matter": "^4.0.3",
